+++ src/sdk/PnP.Core.Admin/Model/SharePoint/Core/Internal/LegacyPrincipalManagement.cs
@@ -43,17 +43,31 @@ namespace PnP.Core.Admin.Model.SharePoint
                 if (loadLegacyPrincipalData)
                 {
                     // Generic principal load, we need this to know the permissions that the AddIns have
-                    await LoadLegacyPrincipalsAsync(legacyPrincipals, null, serverRelativeUrlBuckets, tenantAdminContext).ConfigureAwait(false);
+                    await LoadLegacyPrincipalsAsync(legacyPrincipals, null, serverRelativeUrlBuckets, tenantAdminContext, vanityUrlOptions, context.Uri.DnsSafeHost).ConfigureAwait(false);
                 }
 
                 foreach (var serverRelativeUrlBucket in serverRelativeUrlBuckets)
                 {
-                    var json = new
-                    {
-                        serverRelativeUrls = serverRelativeUrlBucket,
-                    }.AsExpando();
+                    string body = null;
 
-                    var body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                    if (vanityUrlOptions != null)
+                    {
+                        var json = new
+                        {
+                            urls = AsFullyQualified(serverRelativeUrlBucket, context.Uri.DnsSafeHost),
+                        }.AsExpando();
+
+                        body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                    }
+                    else
+                    {
+                        var json = new
+                        {
+                            serverRelativeUrls = serverRelativeUrlBucket,
+                        }.AsExpando();
+
+                        body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                    }
 
                     // Get the AddIns
                     var results = await (tenantAdminContext.Web as Web).RawRequestAsync(new ApiCall($"_api/web/AvailableAddIns", ApiType.SPORest, body), HttpMethod.Post).ConfigureAwait(false);
@@ -103,7 +117,16 @@ namespace PnP.Core.Admin.Model.SharePoint
                             if (loadLegacyPrincipalData)
                             {
                                 // Complement the AddIn with the permission related information
-                                var legacyPrincipal = legacyPrincipals.FirstOrDefault(p => p.AppIdentifier == sharePointAddIn.AppIdentifier && p.ServerRelativeUrl == sharePointAddIn.ServerRelativeUrl);
+                                LegacyPrincipal legacyPrincipal = null;
+                                if (vanityUrlOptions != null)
+                                {
+                                    legacyPrincipal = legacyPrincipals.FirstOrDefault(p => p.AppIdentifier == sharePointAddIn.AppIdentifier && p.AbsoluteUrl == sharePointAddIn.AbsoluteUrl);
+                                }
+                                else
+                                {
+                                    legacyPrincipal = legacyPrincipals.FirstOrDefault(p => p.AppIdentifier == sharePointAddIn.AppIdentifier && p.ServerRelativeUrl == sharePointAddIn.ServerRelativeUrl);
+                                }
+                                
                                 if (legacyPrincipal != null)
                                 {
                                     sharePointAddIn.SiteCollectionScopedPermissions = legacyPrincipal.SiteCollectionScopedPermissions;
@@ -132,7 +155,7 @@ namespace PnP.Core.Admin.Model.SharePoint
                 context.Web.Url.LocalPath
             };
 
-            if (includeSubsites && !tenantOnly) 
+            if (includeSubsites && !tenantOnly)
             {
                 var webs = await WebEnumerator.GetWithDetailsAsync(context, null, true).ConfigureAwait(false);
                 foreach (var web in webs)
@@ -146,7 +169,7 @@ namespace PnP.Core.Admin.Model.SharePoint
             using (var tenantAdminContext = await context.GetSharePointAdmin().GetTenantAdminCenterContextAsync(vanityUrlOptions).ConfigureAwait(false))
             {
                 // Generic principal load, this will get the permission information of the ACS principals
-                await LoadLegacyPrincipalsAsync(legacyPrincipals, legacyServicePrincipals, serverRelativeUrlBuckets, tenantAdminContext).ConfigureAwait(false);
+                await LoadLegacyPrincipalsAsync(legacyPrincipals, legacyServicePrincipals, serverRelativeUrlBuckets, tenantAdminContext, vanityUrlOptions, context.Uri.DnsSafeHost).ConfigureAwait(false);
 
                 // ACS specific data loading
                 List<string> appIds = new();
@@ -215,13 +238,24 @@ namespace PnP.Core.Admin.Model.SharePoint
                             var tempACSPrincipal = tempACSPrincipals.FirstOrDefault(p => p.AppIdentifier == legacyPrincipal.AppIdentifier);
                             if (tempACSPrincipal != null)
                             {
-                                if (!acsPrincipals.Any(p => p.AppIdentifier == legacyPrincipal.AppIdentifier && p.ServerRelativeUrl == legacyPrincipal.ServerRelativeUrl))
+                                bool acsPrincipalFound = false;
+                                if (vanityUrlOptions != null)
+                                {
+                                    acsPrincipalFound = acsPrincipals.Any(p => p.AppIdentifier == legacyPrincipal.AppIdentifier && p.AbsoluteUrl == legacyPrincipal.AbsoluteUrl);
+                                }
+                                else
+                                {
+                                    acsPrincipalFound = acsPrincipals.Any(p => p.AppIdentifier == legacyPrincipal.AppIdentifier && p.ServerRelativeUrl == legacyPrincipal.ServerRelativeUrl);
+                                }
+
+                                if (!acsPrincipalFound)
                                 {
                                     acsPrincipals.Add(new ACSPrincipal
                                     {
                                         AppIdentifier = legacyPrincipal.AppIdentifier,
                                         AllowAppOnly = legacyPrincipal.AllowAppOnly,
                                         ServerRelativeUrl = legacyPrincipal.ServerRelativeUrl,
+                                        AbsoluteUrl = legacyPrincipal.AbsoluteUrl,
                                         SiteCollectionScopedPermissions = legacyPrincipal.SiteCollectionScopedPermissions,
                                         TenantScopedPermissions = legacyPrincipal.TenantScopedPermissions,
                                         Title = legacyPrincipal.Title,
@@ -235,15 +269,27 @@ namespace PnP.Core.Admin.Model.SharePoint
                         }
 
                         // As we've only added principals with tenant permissions for the first site so far we need to copy them for all other investigated webs
-                        foreach (var acsPrincipal in acsPrincipals.ToList().Where(p => p.ServerRelativeUrl == serverRelativeUrlsToCheck[0] && p.TenantScopedPermissions.Any()))
+                        IEnumerable<ACSPrincipal> acsPrincipalsToCopy = null;
+
+                        if (vanityUrlOptions != null)
                         {
-                            for (int i = 1; i < serverRelativeUrlsToCheck.Count; i++) 
-                            { 
+                            acsPrincipalsToCopy = acsPrincipals.ToList().Where(p => p.AbsoluteUrl == $"https://{context.Uri.DnsSafeHost}{serverRelativeUrlsToCheck[0]}" && p.TenantScopedPermissions.Any());
+                        }
+                        else
+                        {
+                            acsPrincipalsToCopy = acsPrincipals.ToList().Where(p => p.ServerRelativeUrl == serverRelativeUrlsToCheck[0] && p.TenantScopedPermissions.Any());
+                        }
+
+                        foreach (var acsPrincipal in acsPrincipalsToCopy)
+                        {
+                            for (int i = 1; i < serverRelativeUrlsToCheck.Count; i++)
+                            {
                                 acsPrincipals.Add(new ACSPrincipal
                                 {
                                     AppIdentifier = acsPrincipal.AppIdentifier,
                                     AllowAppOnly = acsPrincipal.AllowAppOnly,
                                     ServerRelativeUrl = serverRelativeUrlsToCheck[i],
+                                    AbsoluteUrl = $"https://{context.Uri.DnsSafeHost}{serverRelativeUrlsToCheck[i]}",
                                     SiteCollectionScopedPermissions = acsPrincipal.SiteCollectionScopedPermissions,
                                     TenantScopedPermissions = acsPrincipal.TenantScopedPermissions,
                                     Title = acsPrincipal.Title,
@@ -267,9 +313,10 @@ namespace PnP.Core.Admin.Model.SharePoint
                             }
 
                             // empty server relative url as these apply for all server relative urls
-                            foreach(var acsPrincipal in acsPrincipals)
+                            foreach (var acsPrincipal in acsPrincipals)
                             {
                                 acsPrincipal.ServerRelativeUrl = "";
+                                acsPrincipal.AbsoluteUrl = "";
                             }
                         }
                     }
@@ -306,12 +353,12 @@ namespace PnP.Core.Admin.Model.SharePoint
                 // load the valid principals
                 if (jsonResponse.TryGetProperty("value", out JsonElement value) && value.ValueKind == JsonValueKind.Array)
                 {
-                    foreach(var legacyServicePrincipal in value.EnumerateArray())
+                    foreach (var legacyServicePrincipal in value.EnumerateArray())
                     {
                         if (legacyServicePrincipal.TryGetProperty("passwordCredentials", out JsonElement keyCredentials) && keyCredentials.ValueKind == JsonValueKind.Array)
                         {
                             // Only include service principals which are still valid
-                            foreach(var keyCredential in keyCredentials.EnumerateArray())
+                            foreach (var keyCredential in keyCredentials.EnumerateArray())
                             {
                                 var endDate = DateTime.MinValue;
                                 if (keyCredential.TryGetProperty("endDateTime", out JsonElement endDateTime))
@@ -349,19 +396,32 @@ namespace PnP.Core.Admin.Model.SharePoint
             return servicePrincipals;
         }
 
-        private static async Task LoadLegacyPrincipalsAsync(List<LegacyPrincipal> legacyPrincipals, List<ILegacyServicePrincipal> legacyServicePrincipals, List<List<string>> serverRelativeUrlBuckets, PnPContext tenantAdminContext)
+        private static async Task LoadLegacyPrincipalsAsync(List<LegacyPrincipal> legacyPrincipals, List<ILegacyServicePrincipal> legacyServicePrincipals, List<List<string>> serverRelativeUrlBuckets, PnPContext tenantAdminContext, VanityUrlOptions vanityUrlOptions, string dnsSafeHost)
         {
             int bucketCount = 1;
             foreach (var serverRelativeUrlBucket in serverRelativeUrlBuckets)
             {
                 // Step 1: Identify which principals have permissions in the web(s)
-                var json = new
+                string body = null;
+
+                if (vanityUrlOptions != null)
                 {
-                    serverRelativeUrls = serverRelativeUrlBucket,
-                }.AsExpando();
+                    var json = new
+                    {
+                        urls = AsFullyQualified(serverRelativeUrlBucket, dnsSafeHost)
+                    }.AsExpando();
 
-                var body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                    body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                }
+                else
+                {
+                    var json = new
+                    {
+                        serverRelativeUrls = serverRelativeUrlBucket,
+                    }.AsExpando();
 
+                    body = JsonSerializer.Serialize(json, typeof(ExpandoObject));
+                }
                 var results = await (tenantAdminContext.Web as Web).RawRequestAsync(new ApiCall($"_api/web/GetAddinPrincipalsHavingPermissionsInSites", ApiType.SPORest, body), HttpMethod.Post).ConfigureAwait(false);
 
                 var jsonResponse = JsonSerializer.Deserialize<JsonElement>(results.Json);
@@ -375,6 +435,7 @@ namespace PnP.Core.Admin.Model.SharePoint
                             {
                                 AppIdentifier = addInPrincipal.GetProperty("appIdentifier").GetString(),
                                 ServerRelativeUrl = addInPrincipal.GetProperty("serverRelativeUrl").GetString(),
+                                AbsoluteUrl = addInPrincipal.GetProperty("absoluteUrl").ToString(),
                                 Title = addInPrincipal.GetProperty("title").GetString(),
                                 SiteCollectionScopedPermissions = new List<LegacySiteCollectionPermission>(),
                                 TenantScopedPermissions = new List<LegacyTenantPermission>()
@@ -390,12 +451,24 @@ namespace PnP.Core.Admin.Model.SharePoint
                     foreach (var legacyServicePrincipal in legacyServicePrincipals)
                     {
                         // As the list also contains the site collection scoped ACS principals we need to check for duplicates
-                        if (!legacyPrincipals.Any(p => p.AppIdentifier == legacyServicePrincipal.AppIdentifier && p.ServerRelativeUrl == serverRelativeUrlBucket[0]))
+                        bool legacyPrincipalFound = false;
+
+                        if (vanityUrlOptions != null)
+                        {
+                            legacyPrincipalFound = legacyPrincipals.Any(p => p.AppIdentifier == legacyServicePrincipal.AppIdentifier && p.AbsoluteUrl == $"https://{dnsSafeHost}{serverRelativeUrlBucket[0]}");
+                        }
+                        else
+                        {
+                            legacyPrincipalFound = legacyPrincipals.Any(p => p.AppIdentifier == legacyServicePrincipal.AppIdentifier && p.ServerRelativeUrl == serverRelativeUrlBucket[0]);
+                        }
+
+                        if (!legacyPrincipalFound)
                         {
                             legacyPrincipals.Add(new LegacyPrincipal()
                             {
                                 AppIdentifier = legacyServicePrincipal.AppIdentifier,
                                 ServerRelativeUrl = serverRelativeUrlBucket[0],
+                                AbsoluteUrl = $"https://{dnsSafeHost}{serverRelativeUrlBucket[0]}",
                                 Title = legacyServicePrincipal.Name,
                                 SiteCollectionScopedPermissions = new List<LegacySiteCollectionPermission>(),
                                 TenantScopedPermissions = new List<LegacyTenantPermission>()
@@ -417,12 +490,24 @@ namespace PnP.Core.Admin.Model.SharePoint
                     results = new List<string>() { legacyPrincipal.AppIdentifier }
                 }.AsExpando();
 
-                var addin = new
+                if (vanityUrlOptions != null)
                 {
-                    serverRelativeUrl = legacyPrincipal.ServerRelativeUrl,
-                    appIdentifiers,
-                };
-                addinsToQuery.Add(addin.AsExpando());
+                    var addin = new
+                    {
+                        url = legacyPrincipal.AbsoluteUrl,
+                        appIdentifiers,
+                    };
+                    addinsToQuery.Add(addin.AsExpando());
+                }
+                else
+                {
+                    var addin = new
+                    {
+                        serverRelativeUrl = legacyPrincipal.ServerRelativeUrl,
+                        appIdentifiers,
+                    };
+                    addinsToQuery.Add(addin.AsExpando());
+                }
             }
 
             if (legacyServicePrincipals != null)
@@ -435,12 +520,24 @@ namespace PnP.Core.Admin.Model.SharePoint
                         results = new List<string>() { legacyPrincipal.AppIdentifier }
                     }.AsExpando();
 
-                    var addin = new
+                    if (vanityUrlOptions != null)
                     {
-                        serverRelativeUrl = serverRelativeUrlBuckets[0][0],
-                        appIdentifiers,
-                    };
-                    addinsToQuery.Add(addin.AsExpando());
+                        var addin = new
+                        {
+                            url = $"https://{dnsSafeHost}{serverRelativeUrlBuckets[0][0]}",
+                            appIdentifiers,
+                        };
+                        addinsToQuery.Add(addin.AsExpando());
+                    }
+                    else
+                    {
+                        var addin = new
+                        {
+                            serverRelativeUrl = serverRelativeUrlBuckets[0][0],
+                            appIdentifiers,
+                        };
+                        addinsToQuery.Add(addin.AsExpando());
+                    }
                 }
             }
 
@@ -464,10 +561,18 @@ namespace PnP.Core.Admin.Model.SharePoint
                     {
                         foreach (var addInPermission in addinPermissions.EnumerateArray())
                         {
-                            string appIdentifier = addInPermission.GetProperty("appIdentifier").GetString();
-                            string serverRelativeUrl = addInPermission.GetProperty("serverRelativeUrl").GetString();
-
-                            var legacyPrincipalToUpdate = legacyPrincipals.First(p => p.AppIdentifier == appIdentifier && p.ServerRelativeUrl == serverRelativeUrl);
+                            string appIdentifier = addInPermission.GetProperty("appIdentifier").GetString();                            
+                            LegacyPrincipal legacyPrincipalToUpdate = null;
+                            if (vanityUrlOptions != null)
+                            {
+                                string absoluteUrl = addInPermission.GetProperty("absoluteUrl").GetString();
+                                legacyPrincipalToUpdate = legacyPrincipals.First(p => p.AppIdentifier == appIdentifier && p.AbsoluteUrl == absoluteUrl);
+                            }
+                            else
+                            {
+                                string serverRelativeUrl = addInPermission.GetProperty("serverRelativeUrl").GetString();
+                                legacyPrincipalToUpdate = legacyPrincipals.First(p => p.AppIdentifier == appIdentifier && p.ServerRelativeUrl == serverRelativeUrl);
+                            }
 
                             legacyPrincipalToUpdate.AllowAppOnly = addInPermission.GetProperty("allowAppOnly").GetBoolean();
 
@@ -519,8 +624,8 @@ namespace PnP.Core.Admin.Model.SharePoint
         private static string AppIdFromAppIdentifier(string appIdentifier)
         {
             var split1 = appIdentifier.Split('@');
-            if (split1.Length == 2) 
-            { 
+            if (split1.Length == 2)
+            {
                 var split2 = split1[0].Split('|');
                 if (split2.Length == 3)
                 {
@@ -560,6 +665,18 @@ namespace PnP.Core.Admin.Model.SharePoint
             }
 
             return splitList;
+        }
+
+        private static List<string> AsFullyQualified(List<string> serverRelativeUrlList, string dnsSafeHost)
+        {
+            List<string> fullyQualified = new();
+
+            for (int i = 0; i < serverRelativeUrlList.Count; i++)
+            {
+                fullyQualified.Add($"https://{dnsSafeHost}{serverRelativeUrlList[i]}");
+            }
+
+            return fullyQualified;
         }
     }
 }