+++ src/sdk/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - `IContentType.SchemaXml` is not settable in SharePoint so removing the set option for it #1418 [jansenbe - Bert Jansen]
 - Ensure `GetFolderByServerRelativeUrl` and `GetFileByServerRelativeUr	l` can handle \ as path separator #1412 [jansenbe - Bert Jansen]
 - GetComments() with a selector only return maximum of 30 comments #1424 [jansenbe - Bert Jansen]
+- Admin library: Improved support for vanity URL tenants when loading ACS principals and SharePoint Add-Ins [jansenbe - Bert Jansen]
 
 ## [1.12.0]